pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "outheis"
version = "0.2.0"
description = "Multi-agent personal assistant"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
authors = [
    { name = "outheis-labs" },
]
keywords = [
    "ai",
    "assistant",
    "multi-agent",
    "llm",
    "personal-assistant",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.100.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]
signal = [
    "faster-whisper>=1.0.0",  # Voice transcription (optional, CPU-only)
]
watch = [
    "watchdog>=4.0",
]

[project.scripts]
outheis = "outheis.cli.main:app"

[project.urls]
Homepage = "https://github.com/outheis-labs/outheis-minimal"
Documentation = "https://outheis-labs.github.io/outheis-minimal/"
Repository = "https://github.com/outheis-labs/outheis-minimal"
Issues = "https://github.com/outheis-labs/outheis-minimal/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/outheis"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=outheis --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (require API key)",
]
